requirements.txt
# CareerAIpilot Dependencies

# Core AI Frameworks
crewai>=0.28.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langgraph>=0.0.40

# LLM Providers
langchain-google-genai>=1.0.0
openai>=1.0.0

# Vector Databases and Embeddings
chromadb>=0.4.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0

# Document Processing
PyPDF2>=3.0.0
PyMuPDF>=1.23.0
python-docx>=0.8.11
docx2txt>=0.8
reportlab>=4.0.0

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Utilities
python-dotenv>=1.0.0
numpy>=1.24.0
pydantic>=2.0.0
typing-extensions>=4.8.0
